better autosampling
